+++ activerecord/lib/active_record/validations.rb
@@ -76,4 +76,5 @@ end
 require "active_record/validations/associated"
 require "active_record/validations/uniqueness"
 require "active_record/validations/presence"
+require "active_record/validations/absence"
 require "active_record/validations/length"