+++ activerecord/lib/active_record/gem_version.rb
@@ -7,8 +7,8 @@ module ActiveRecord
   module VERSION
     MAJOR = 4
     MINOR = 2
-    TINY  = 2
-    PRE   = nil
+    TINY  = 3
+    PRE   = "rc1"
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")
   end